git-svn-id: https://unicode.org/repos/unicodetools/trunk@1382 13e8329f-0b23-4da4-9fe8-d0f6fe080806

+++ unicodetools/org/unicode/text/UCD/UData.java
@@ -483,7 +483,7 @@ class UData implements UCD_Types {
         os.writeByte(eastAsianWidth);
         os.writeByte(lineBreak);
         os.writeByte(joiningType);
-        os.writeByte(joiningGroup);
+        os.writeShort(joiningGroup);
         os.writeShort(script);
         os.writeByte(age);
         os.writeByte(Bidi_Paired_Bracket_Type);
@@ -518,7 +518,7 @@ class UData implements UCD_Types {
         eastAsianWidth = is.readByte();
         lineBreak = is.readByte();
         joiningType = is.readByte();
-        joiningGroup = is.readByte();
+        joiningGroup = is.readShort();
         script = is.readShort();
         age = is.readByte();
         Bidi_Paired_Bracket_Type = is.readByte();
